pyproject.toml
[tool.poetry]
name = "kumo"
version = "0.1.0"
description = "Kumo is a vulnerability management tool for container images."
authors = ["2shiori17 <98276492+2shiori17@users.noreply.github.com>"]
license = "MPL-2.0"
readme = "README.md"

[tool.poetry.dependencies]
apache-beam = "^2.53.0"
fastapi = {extras = ["all"], version = "^0.109.0"}
gqlalchemy = {extras = ["arrow", "docker"], version = "^1.5.1"}
pymgclient = "^1.3.1"
python = "^3.11"
strawberry-graphql = {extras = ["fastapi"], version = "^0.217.1"}
uvicorn = {extras = ["standard"], version = "^0.25.0"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
ruff = "^0.1.13"
taskipy = "^1.12.2"
commitizen = "^3.13.0"
pre-commit = "^3.6.0"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
check = "task check_mypy && task check_ruff"
check_mypy = "mypy ."
check_ruff = "ruff check"
fmt = "ruff check --fix"
test = "pytest --cov"
dev_server = "uvicorn kumo_server:api --reload"

[tool.mypy]
python_version = "3.11"
plugins = "strawberry.ext.mypy_plugin"

[tool.pyright]
reportPrivateImportUsage = false
